# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: ESG ProofOps API
  version: 1.0.0
  description: Implementation contract, not a deployed service. Domain v2.0 is authoritative.
servers:
- url: /
paths:
...
  /v1/runs/{run_id}/reviews:
    get:
      operationId: reviews_list
      summary: reviews_list
      x-minimum-role: reviewer
      x-rate-limit: 120/min/user
      x-idempotency-required: false
      security:
      - sessionCookie: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewPage'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden or CSRF failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found or tenant mismatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Conflict / gate blocked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '412':
          description: Stale revision
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Schema or domain validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate/budget limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Dependency unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
      - name: run_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: cursor
        in: query
        schema:
          type: string
      - name: limit
        in: query
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 50
  /v1/reviews/{review_id}/resolve:
    post:
      operationId: review_resolve
      summary: review_resolve
      x-minimum-role: reviewer
      x-rate-limit: 10/min/user
      x-idempotency-required: true
      security:
      - sessionCookie: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewResolution'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden or CSRF failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found or tenant mismatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Conflict / gate blocked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '412':
          description: Stale revision
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Schema or domain validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate/budget limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Dependency unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      parameters:
      - name: review_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-CSRF-Token
        in: header
        required: true
        schema:
          type: string
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          minLength: 16
          maxLength: 128
      - name: If-Match
        in: header
        required: true
        schema:
          type: string
          pattern: ^"[1-9][0-9]*"$
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewResolve'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-proofops_session
  schemas:
    Error:
      type: object
      additionalProperties: false
      properties:
        error:
          type: object
          additionalProperties: false
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
              format: uuid
            retryable: &id003
              type: boolean
            details:
              type: object
              additionalProperties: false
              properties:
                field:
                  type: string
                reason:
                  type: string
                current_revision:
                  type: integer
                  minimum: 1
                retry_after_seconds:
                  type: integer
                  minimum: 0
          required:
          - code
          - message
          - request_id
          - retryable
      required:
      - error
...
    Bbox:
      type: array
      items:
        type: number
      minItems: 4
      maxItems: 4
      description: Canonical top-left display points, x0,y0,x1,y1; missing geometry is null, not zero.
    SourceRef:
      type: object
      additionalProperties: false
      properties:
        source_id:
          type: string
          format: uuid
        document_version_id:
          type: string
          format: uuid
        parse_manifest_id:
          type: string
          format: uuid
        page_num:
          type: integer
          minimum: 1
        printed_page_label:
          anyOf:
          - type: string
          - type: 'null'
        bbox:
          anyOf:
          - $ref: '#/components/schemas/Bbox'
          - type: 'null'
        raw_text_sha256: *id002
        quote:
          type: string
        char_start:
          type: integer
          minimum: 0
        char_end:
          type: integer
          minimum: 0
        location_quality:
          type: string
          enum:
          - located
          - unlocated
          - unreadable
        verification_state:
          type: string
          enum:
          - verified
          - candidate
          - rejected
      required:
      - source_id
      - document_version_id
      - parse_manifest_id
      - page_num
      - printed_page_label
      - bbox
      - raw_text_sha256
      - quote
      - char_start
      - char_end
      - location_quality
      - verification_state
    Element:
      type: object
      additionalProperties: false
      properties:
        element_id:
          type: string
        state:
          type: string
          enum:
          - present
          - absent
          - unknown
          - conflict
          - not_applicable
        evidence_refs:
          type: array
          items:
            $ref: '#/components/schemas/SourceRef'
        normalized_value:
          anyOf:
          - type: string
          - type: 'null'
        credited_from:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
        reason_code:
          anyOf:
          - type: string
          - type: 'null'
      required:
      - element_id
      - state
      - evidence_refs
      - normalized_value
      - credited_from
      - reason_code
      allOf:
      - if:
          properties:
            state:
              const: present
        then:
          properties:
            evidence_refs:
              minItems: 1
    Decision:
      type: object
      additionalProperties: false
      properties:
        decision_revision:
          type: integer
          minimum: 1
        tag_revision:
          type: integer
          minimum: 1
        decision_status:
          type: string
          enum:
          - decided
          - blocked_evidence
          - blocked_rule_gap
          - not_applicable
          - not_run
        evidence_grade:
          anyOf:
          - type: string
            enum:
            - E0
            - E1
            - E2
            - E3
          - type: 'null'
        label:
          anyOf:
          - type: string
            enum:
            - SUBSTANTIATED
            - INCOMPLETE
            - UNSUBSTANTIATED
          - type: 'null'
        sublabel:
          anyOf:
          - type: string
            enum:
            - PERF
            - IMPL
          - type: 'null'
        review_status:
          type: string
          enum:
          - auto_confirmed
          - needs_review
          - human_confirmed
        missing_elements:
          type: array
          items:
            type: string
        rule_ids:
          type: array
          items:
            type: string
        rule_pack_sha256: *id002
        semantic_hash: *id002
        gap_ids:
          type: array
          items:
            type: string
      required:
      - decision_revision
      - tag_revision
      - decision_status
      - evidence_grade
      - label
      - sublabel
      - review_status
      - missing_elements
      - rule_ids
      - rule_pack_sha256
      - semantic_hash
      - gap_ids
      allOf:
      - if:
          properties:
            decision_status:
              const: decided
        then:
          properties:
            evidence_grade:
              type: string
              enum:
              - E0
              - E1
              - E2
              - E3
            label:
              type: string
              enum:
              - SUBSTANTIATED
              - INCOMPLETE
              - UNSUBSTANTIATED
        else:
          properties:
            evidence_grade:
              type: 'null'
            label:
              type: 'null'
            sublabel:
              type: 'null'
      - if:
          properties:
            evidence_grade:
              const: E0
        then:
          properties:
            label:
              const: UNSUBSTANTIATED
      - if:
          properties:
            evidence_grade:
              const: E1
        then:
          properties:
            label:
              const: INCOMPLETE
      - if:
          properties:
            evidence_grade:
              const: E2
        then:
          properties:
            label:
              const: INCOMPLETE
      - if:
          properties:
            evidence_grade:
              const: E3
        then:
          properties:
            label:
              const: SUBSTANTIATED
...
    Review:
      type: object
      additionalProperties: false
      properties:
        review_id:
          type: string
          format: uuid
        run_id:
          type: string
          format: uuid
        claim_id:
          type: string
          format: uuid
        status:
          type: string
          enum:
          - open
          - resolved
          - superseded
        revision:
          type: integer
          minimum: 1
        base_tag_revision:
          type: integer
          minimum: 1
        reason_codes:
          type: array
          items:
            type: string
      required:
      - review_id
      - run_id
      - claim_id
      - status
      - revision
      - base_tag_revision
      - reason_codes
    ReviewResolve:
      type: object
      additionalProperties: false
      properties:
        base_tag_revision:
          type: integer
          minimum: 1
        track:
          type: string
          enum:
          - goal
          - performance
          - management
        elements:
          type: array
          items:
            $ref: '#/components/schemas/Element'
        reason:
          type: string
          minLength: 5
          maxLength: 1000
      required:
      - base_tag_revision
      - track
      - elements
      - reason
    ReviewResolution:
      type: object
      additionalProperties: false
      properties:
        review:
          $ref: '#/components/schemas/Review'
        decision:
          $ref: '#/components/schemas/Decision'
        new_tag_revision:
          type: integer
          minimum: 1
      required:
      - review
      - decision
      - new_tag_revision
...
    ReviewPage:
      type: object
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Review'
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
        snapshot_epoch:
          anyOf:
          - type: integer
            minimum: 0
          - type: 'null'
      required:
      - items
      - next_cursor
      - snapshot_epoch
...